+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.6.28] - 2021-07-03
 ### Added
 - Enchantment description for Star Charged (TomasBorsje) [#399, #388]
 - Fireproof trait: makes it so item cannot be destroyed in fire or lava. Flame Ward also has the same effect. [#309]